:recycle: Refactor the backend file structures
