+++ src/Grisette/Backend/SymBiMap.hs
@@ -1,14 +1,14 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 
 -- |
--- Module      :   Grisette.Backend.SBV.Data.SMT.SymBiMap
+-- Module      :   Grisette.Backend.SymBiMap
 -- Copyright   :   (c) Sirui Lu 2021-2023
 -- License     :   BSD-3-Clause (see the LICENSE file)
 --
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.Backend.SBV.Data.SMT.SymBiMap
+module Grisette.Backend.SymBiMap
   ( SymBiMap (..),
     emptySymBiMap,
     sizeBiMap,
