+++ src/Grisette/Core/Data/Class/Solver.hs
@@ -80,7 +80,7 @@ data SolveInternal = SolveInternal
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> :set -XOverloadedStrings
 
 -- | The current failures that can be returned by the solver.
