+++ src/Grisette/Internal/Backend.hs
@@ -1,19 +1,19 @@
 -- |
--- Module      :   Grisette.Internal.Backend.SBV
+-- Module      :   Grisette.Internal.Backend
 -- Copyright   :   (c) Sirui Lu 2021-2023
 -- License     :   BSD-3-Clause (see the LICENSE file)
 --
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.Internal.Backend.SBV
+module Grisette.Internal.Backend
   ( lowerSinglePrim,
     lowerSinglePrimCached,
     parseModel,
   )
 where
 
-import Grisette.Backend.SBV.Data.SMT.Solving
+import Grisette.Backend.Solving
   ( lowerSinglePrim,
     lowerSinglePrimCached,
     parseModel,
