+++ src/Grisette/SymPrim/AllSyms.hs
@@ -72,7 +72,7 @@ import Grisette.SymPrim.Prim.TermUtils
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> import Data.Proxy
 
 -- | Get the sum of the sizes of a list of symbolic terms.
