+++ src/Grisette/SymPrim/ModelRep.hs
@@ -24,7 +24,7 @@ import Grisette.SymPrim.Prim.Term
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> import Data.Proxy
 
 -- ModelRep
