+++ src/Grisette/SymPrim/SymBV.hs
@@ -137,7 +137,7 @@ import Language.Haskell.TH.Syntax (Lift)
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> import Data.Proxy
 
 -- | Symbolic signed bit vector type. Indexed with the bit width.
