+++ src/Grisette/SymPrim/SymBool.hs
@@ -35,7 +35,7 @@ import Language.Haskell.TH.Syntax (Lift)
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> import Data.Proxy
 
 -- | Symbolic Boolean type.
