+++ src/Grisette/SymPrim/SymInteger.hs
@@ -43,7 +43,7 @@ import Language.Haskell.TH.Syntax (Lift)
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> import Data.Proxy
 
 -- | Symbolic (unbounded, mathematical) integer type.
