+++ src/Grisette/SymPrim/SymTabularFun.hs
@@ -48,7 +48,7 @@ import Language.Haskell.TH.Syntax (Lift (liftTyped))
 -- $setup
 -- >>> import Grisette.Core
 -- >>> import Grisette.SymPrim
--- >>> import Grisette.Backend.SBV
+-- >>> import Grisette.Backend
 -- >>> import Data.Proxy
 
 -- | Symbolic tabular function type.
